clean up .legacy files
